+++ voucher/src/main/java/org/prgms/voucher/application/Console.java
@@ -1,4 +1,4 @@
-package org.prgms.voucher;
+package org.prgms.voucher.application;
 
 public interface Console {
     void printSupportedCommands();
