+++ voucher/src/main/java/org/prgms/voucher/voucher/Voucher.java
@@ -1,4 +1,4 @@
-package org.prgms.voucher;
+package org.prgms.voucher.voucher;
 
 public interface Voucher {
     void discount();
