+++ voucher/src/main/java/org/prgms/voucher/voucher/repository/MemoryVoucherRepository.java
@@ -1,4 +1,7 @@
-package org.prgms.voucher;
+package org.prgms.voucher.voucher.repository;
+
+import org.prgms.voucher.voucher.Voucher;
+import org.prgms.voucher.voucher.repository.VoucherRepository;
 
 import java.util.List;
